third_party/blink/web_tests/wpt_internal/prerender/script-supports-speculationrules.html

<!DOCTYPE html>
<html>
<title>HTMLScriptElement.supports speculationrules</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="timeout" content="long">
<script>
test(function() {
  assert_true(HTMLScriptElement.supports('speculationrules'));
}, 'HTMLScriptElement.supports returns true for \'speculationrules\'');

test(function() {
  assert_false(HTMLScriptElement.supports(' speculationrules'));
  assert_false(HTMLScriptElement.supports('speculationrules '));
  assert_false(HTMLScriptElement.supports('Speculationrules'));
  assert_false(HTMLScriptElement.supports('SpeculationRules'));
  assert_false(HTMLScriptElement.supports('speculationRules'));
  assert_false(HTMLScriptElement.supports('speculation-rules'));
}, 'HTMLScriptElement.supports returns false for unsupported types');

</script>
